CR-681 RHUI Unlinked UAC Addressed some review issues - moved InvalidDataError and InvalidDataErrorWelsh
