+++ app/utils.py
@@ -2,8 +2,7 @@ import string
 import re
 import json
 
-from .exceptions import InactiveCaseError
-from .exceptions import InvalidEqPayLoad
+from .exceptions import InactiveCaseError, InvalidEqPayLoad, InvalidDataError, InvalidDataErrorWelsh
 from aiohttp.web import HTTPFound
 from datetime import datetime, timezone
 
@@ -99,18 +98,6 @@ class View:
                                         json=launch_json)
 
 
-class InvalidDataError(Exception):
-
-    def __init__(self, message=None):
-        super().__init__(message or 'The supplied value is invalid')
-
-
-class InvalidDataErrorWelsh(Exception):
-
-    def __init__(self, message=None):
-        super().__init__(message or 'WELSH The supplied value is invalid')
-
-
 class ProcessPostcode:
     postcode_validation_pattern = re.compile(
         r'^((AB|AL|B|BA|BB|BD|BH|BL|BN|BR|BS|BT|BX|CA|CB|CF|CH|CM|CO|CR|CT|CV|CW|DA|DD|DE|DG|DH|DL|DN|DT|DY|E|EC|EH|EN|EX|FK|FY|G|GL|GY|GU|HA|HD|HG|HP|HR|HS|HU|HX|IG|IM|IP|IV|JE|KA|KT|KW|KY|L|LA|LD|LE|LL|LN|LS|LU|M|ME|MK|ML|N|NE|NG|NN|NP|NR|NW|OL|OX|PA|PE|PH|PL|PO|PR|RG|RH|RM|S|SA|SE|SG|SK|SL|SM|SN|SO|SP|SR|SS|ST|SW|SY|TA|TD|TF|TN|TQ|TR|TS|TW|UB|W|WA|WC|WD|WF|WN|WR|WS|WV|YO|ZE)(\d[\dA-Z]?[ ]?\d[ABD-HJLN-UW-Z]{2}))|BFPO[ ]?\d{1,4}$'  # NOQA
@@ -126,30 +113,35 @@ class ProcessPostcode:
 
         if len(postcode) == 0:
             if locale == 'cy':
+                # TODO: Add Welsh Translation
                 raise InvalidDataErrorWelsh('You have not entered a postcode')
             else:
                 raise InvalidDataError('You have not entered a postcode')
 
         if not postcode.isalnum():
             if locale == 'cy':
+                # TODO: Add Welsh Translation
                 raise InvalidDataErrorWelsh('The postcode must not contain symbols')
             else:
                 raise InvalidDataError('The postcode must not contain symbols')
 
         if len(postcode) < 5:
             if locale == 'cy':
+                # TODO: Add Welsh Translation
                 raise InvalidDataErrorWelsh('The postcode does not contain enough characters')
             else:
                 raise InvalidDataError('The postcode does not contain enough characters')
 
         if len(postcode) > 7:
             if locale == 'cy':
+                # TODO: Add Welsh Translation
                 raise InvalidDataErrorWelsh('The postcode contains too many characters')
             else:
                 raise InvalidDataError('The postcode contains too many characters')
 
         if not ProcessPostcode.postcode_validation_pattern.fullmatch(postcode):
             if locale == 'cy':
+                # TODO: Add Welsh Translation
                 raise InvalidDataErrorWelsh('The postcode is not a valid UK postcode')
             else:
                 raise InvalidDataError('The postcode is not a valid UK postcode')
@@ -171,6 +163,7 @@ class ProcessMobileNumber:
             list(map(int, number))
         except ValueError:
             if locale == 'cy':
+                # TODO: Add Welsh Translation
                 raise InvalidDataErrorWelsh('The mobile phone number must not contain letters or symbols')
             else:
                 raise InvalidDataError('The mobile phone number must not contain letters or symbols')
@@ -184,18 +177,21 @@ class ProcessMobileNumber:
 
         if not number.startswith('7'):
             if locale == 'cy':
+                # TODO: Add Welsh Translation
                 raise InvalidDataErrorWelsh('The mobile phone number is not a UK mobile number')
             else:
                 raise InvalidDataError('The mobile phone number is not a UK mobile number')
 
         if len(number) > 10:
             if locale == 'cy':
+                # TODO: Add Welsh Translation
                 raise InvalidDataErrorWelsh('The mobile phone number contains too many digits')
             else:
                 raise InvalidDataError('The mobile phone number contains too many digits')
 
         if len(number) < 10:
             if locale == 'cy':
+                # TODO: Add Welsh Translation
                 raise InvalidDataErrorWelsh('The mobile phone number does not contain enough digits')
             else:
                 raise InvalidDataError('The mobile phone number does not contain enough digits')
@@ -220,6 +216,7 @@ class AddressIndex(View):
         address_options = []
 
         if display_region == 'cy':
+            # TODO: Add Welsh Translation
             cannot_find_text = 'I cannot find my address'
         else:
             cannot_find_text = 'I cannot find my address'
